refactor logger to be usable in both node and browser
