+++ src/logger/browserLogger.js
@@ -1,0 +1,1 @@
+export const browserLogger = console;
