+++ src/logger/nodeLogger.js
@@ -1,12 +1,14 @@
 import process from "node:process";
 import winston from "winston";
+import { serialize } from "./serialize.js";
+
 const { combine, timestamp, printf } = winston.format;
 
 const myFormat = printf(({ level, message, timestamp }) => {
   return `${timestamp} [${level}] ${message}`;
 });
 
-export const logger = winston.createLogger({
+export const nodeBrowser = winston.createLogger({
   levels: {
     error: 0,
     warn: 1,
@@ -32,7 +34,7 @@ export const logger = winston.createLogger({
 process.on(
   "unhandledRejection",
   function handleUnhandledRejection(reason, promise) {
-    logger.error(
+    nodeBrowser.error(
       `Unhandled Rejection at: ${serialize(promise)}, reason: ${serialize(
         reason,
       )}`,
@@ -41,23 +43,8 @@ process.on(
 );
 
 process.on("uncaughtException", function handleUncaughtException(error) {
-  logger.error(`There was an uncaught exception: ${serialize(error)}`);
-  logger.on("finish", () => {
+  nodeBrowser.error(`There was an uncaught exception: ${serialize(error)}`);
+  nodeBrowser.on("finish", () => {
     process.exit(1);
   });
 });
-
-export function serialize(something) {
-  if (something instanceof Date) {
-    return something.toISOString();
-  } else if (something instanceof Error) {
-    return something.toString();
-  } else if (typeof something === "function") {
-    return something.toString();
-  } else if (something instanceof Promise) {
-    return "Can't serialize a promise";
-  } else if (something instanceof Response) {
-    return `Request failed with ${something.status} (${something.statusText}). URL: ${something.url}`;
-  }
-  return JSON.stringify(something);
-}
